Move files around to remove mlwh.
